pyproject.toml
[project]
name = "caltech2020-glass-sims"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pybaselines>=1.2.0",
    "pyjanitor>=0.31.0",
    "scipy>=1.15.2",
    "xarray>=2025.3.1",
]
